chore(formatter): apply

+++ contracts/CJPY.sol
@@ -4,7 +4,7 @@ pragma solidity 0.7.6;
  * SPDX-License-Identifier: GPL-3.0-or-later
  * Written by 0xMotoko (0xmotoko@pm.me)
  * Copyright (C) 2021 Yamato Protocol (DeFiGeek Community Japan)
-*/
+ */
 
 //solhint-disable max-line-length
 //solhint-disable no-inline-assembly
@@ -18,9 +18,9 @@ import "./Yamato.sol";
  * @notice Very stable.
  */
 contract CJPY is Currency {
-
-    constructor(uint initialSupply) Currency("Convertible JPY Token","CJPY") {
+    constructor(uint256 initialSupply)
+        Currency("Convertible JPY Token", "CJPY")
+    {
         _mint(msg.sender, initialSupply);
     }
-
 }
