+++ contracts/OracleMockBase.sol
@@ -4,7 +4,7 @@ pragma solidity 0.7.6;
  * SPDX-License-Identifier: GPL-3.0-or-later
  * Written by somewherecat
  * Copyright (C) 2021 Yamato Protocol (DeFiGeek Community Japan)
-*/
+ */
 
 //solhint-disable max-line-length
 //solhint-disable no-inline-assembly
@@ -13,28 +13,35 @@ import "./Dependencies/Ownable.sol";
 
 // Base class to create a oracle mock contract for a specific provider
 abstract contract OracleMockBase is Ownable {
-
     int256 internal lastPrice;
-    uint private lastBlockNumber;
+    uint256 private lastBlockNumber;
 
     function setLastPrice(int256 _price) public onlyOwner {
-      lastPrice = _price;
-      lastBlockNumber = block.number;
+        lastPrice = _price;
+        lastBlockNumber = block.number;
     }
 
     function setPriceToDefault() public virtual;
 
-    function simulatePriceMove(uint deviation, bool sign) internal virtual;
+    function simulatePriceMove(uint256 deviation, bool sign) internal virtual;
 
     function simulatePriceMove() public onlyOwner {
-      // Within each block, only once price update is allowed (volatility control)
-      if (block.number != lastBlockNumber) {
-        lastBlockNumber = block.number;
+        // Within each block, only once price update is allowed (volatility control)
+        if (block.number != lastBlockNumber) {
+            lastBlockNumber = block.number;
 
-        uint randomNumber = uint(keccak256(abi.encodePacked(msg.sender,  block.timestamp,  blockhash(block.number - 1))));
-        uint deviation = randomNumber % 11;
-        bool sign = randomNumber % 2 == 1 ? true : false;
-        simulatePriceMove(deviation, sign);
-      }
+            uint256 randomNumber = uint256(
+                keccak256(
+                    abi.encodePacked(
+                        msg.sender,
+                        block.timestamp,
+                        blockhash(block.number - 1)
+                    )
+                )
+            );
+            uint256 deviation = randomNumber % 11;
+            bool sign = randomNumber % 2 == 1 ? true : false;
+            simulatePriceMove(deviation, sign);
+        }
     }
 }
